+++ src/pages/InvitedPage.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import PageContainer from 'components/@commons/PageContainer';
-import JoinDone from 'components/alba-invited/JoinDone';
-import InvitationSection from 'components/alba-invited/InvitationSection';
+import JoinDone from 'components/alba-InvitedPage/JoinDone';
+import InvitationSection from 'components/alba-InvitedPage/InvitationSection';
 
 const InvitedPage = (): JSX.Element => {
   // 로그인상태인 유저가 이미 그룹에 소속된 경우 : 리다이렉트 "/" 또는 에러페이지
